# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Al Jood Travel & Tourism API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Destination {
  id: number;
  slug: string;
  nameAr: string;
  nameEn: string;
  country: string;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  descriptionEn?: string | null;
  heroImage: string;
  /** @nullable */
  flag?: string | null;
  isActive: boolean;
  isFeatured: boolean;
  sortOrder: number;
  /** @nullable */
  ticketPriceJod?: number | null;
  /** @nullable */
  hotelCount?: number | null;
  /** @nullable */
  minPrice?: number | null;
}

export interface DestinationInput {
  slug: string;
  nameAr: string;
  nameEn: string;
  country: string;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  descriptionEn?: string | null;
  heroImage: string;
  /** @nullable */
  flag?: string | null;
  isActive?: boolean;
  isFeatured?: boolean;
  sortOrder?: number;
  /** @nullable */
  ticketPriceJod?: number | null;
}

export interface DestinationSummary {
  slug: string;
  totalHotels: number;
  totalPackages: number;
  /** @nullable */
  minPrice: number | null;
  /** @nullable */
  maxPrice: number | null;
  availableNights: number[];
  availableAreas: string[];
  availableMealPlans: string[];
  availableStars: number[];
}

export interface Hotel {
  id: number;
  destinationId: number;
  nameAr: string;
  nameEn: string;
  stars: number;
  /** @nullable */
  area: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
}

export interface HotelInput {
  destinationId: number;
  nameAr: string;
  nameEn: string;
  stars: number;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}

export interface Package {
  id: number;
  hotelId: number;
  destinationId: number;
  nights: number;
  mealPlan: string;
  roomType: string;
  basePriceUsd: number;
  currency: string;
  /** @nullable */
  dateFrom?: string | null;
  /** @nullable */
  dateTo?: string | null;
  isActive: boolean;
  /** @nullable */
  hotelNameAr?: string | null;
  /** @nullable */
  hotelNameEn?: string | null;
  /** @nullable */
  hotelStars?: number | null;
  /** @nullable */
  hotelArea?: string | null;
}

export interface PackageInput {
  hotelId: number;
  destinationId: number;
  nights: number;
  mealPlan: string;
  roomType: string;
  basePriceUsd: number;
  currency: string;
  /** @nullable */
  dateFrom?: string | null;
  /** @nullable */
  dateTo?: string | null;
  isActive?: boolean;
}

export interface PackageToggle {
  isActive: boolean;
}

export interface PackageResult {
  id: number;
  hotelId: number;
  hotelNameAr: string;
  hotelNameEn: string;
  stars: number;
  /** @nullable */
  area: string | null;
  mealPlan: string;
  roomType: string;
  nights: number;
  finalPriceJod: number;
  /** @nullable */
  finalPriceUsd?: number | null;
  /** @nullable */
  dateFrom?: string | null;
  /** @nullable */
  dateTo?: string | null;
  isActive: boolean;
  /** @nullable */
  hotelImageUrl?: string | null;
  /** @nullable */
  hotelDescription?: string | null;
}

export interface PricingSettings {
  id: number;
  ticketPriceJod: number;
  transportJod: number;
  fixedProfitJod: number;
  profitPct: number;
  rateUsdToJod: number;
  rateEurToJod: number;
  rateSarToJod: number;
  whatsappNumber: string;
  updatedAt: string;
}

export interface PricingSettingsInput {
  ticketPriceJod?: number;
  transportJod?: number;
  fixedProfitJod?: number;
  profitPct?: number;
  rateUsdToJod?: number;
  rateEurToJod?: number;
  rateSarToJod?: number;
  whatsappNumber?: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminSession {
  id: number;
  username: string;
  role: string;
}

export type AdminStatsDestinationsItem = {
  slug: string;
  nameEn: string;
  hotelCount: number;
  packageCount: number;
};

export interface AdminStats {
  totalDestinations: number;
  totalHotels: number;
  totalPackages: number;
  activePackages: number;
  destinations: AdminStatsDestinationsItem[];
}

export interface QuoteRequest {
  packageId: number;
  nights: number;
  roomType: string;
  guestName: string;
  /** @nullable */
  guestPhone?: string | null;
  adults?: number;
  children?: number;
  /** @nullable */
  specialRequests?: string | null;
}

export interface QuoteResponse {
  quoteId: string;
  hotelNameAr: string;
  hotelNameEn: string;
  destinationNameAr: string;
  destinationNameEn: string;
  nights: number;
  roomType: string;
  mealPlan: string;
  finalPriceJod: number;
  /** @nullable */
  finalPriceUsd?: number | null;
  whatsappNumber: string;
  whatsappMessage: string;
  guestName: string;
  /** @nullable */
  adults?: number | null;
  /** @nullable */
  children?: number | null;
  /** @nullable */
  hotelStars?: number | null;
  /** @nullable */
  hotelArea?: string | null;
  /** @nullable */
  dateFrom?: string | null;
  /** @nullable */
  dateTo?: string | null;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListPackagesParams = {
/**
 * @nullable
 */
destinationSlug?: string | null;
/**
 * @nullable
 */
nights?: number | null;
/**
 * @nullable
 */
stars?: number | null;
/**
 * @nullable
 */
area?: string | null;
/**
 * @nullable
 */
mealPlan?: string | null;
/**
 * @nullable
 */
roomType?: string | null;
/**
 * @nullable
 */
maxPrice?: number | null;
/**
 * @nullable
 */
minPrice?: number | null;
};

export type ListAdminHotelsParams = {
/**
 * @nullable
 */
destinationId?: number | null;
};

export type ListAdminPackagesParams = {
/**
 * @nullable
 */
destinationId?: number | null;
/**
 * @nullable
 */
hotelId?: number | null;
};