# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Al Jood Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /destinations/{slug}:
    get:
      operationId: getDestination
      tags: [destinations]
      summary: Get destination details
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Destination details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
        "404":
          description: Not found
  /packages:
    get:
      operationId: listPackages
      tags: [destinations]
      summary: List computed packages with final prices (filter by destination, nights, stars, etc.)
      parameters:
        - name: destinationSlug
          in: query
          schema:
            type: ["string", "null"]
        - name: nights
          in: query
          schema:
            type: ["integer", "null"]
        - name: stars
          in: query
          schema:
            type: ["integer", "null"]
        - name: area
          in: query
          schema:
            type: ["string", "null"]
        - name: mealPlan
          in: query
          schema:
            type: ["string", "null"]
        - name: roomType
          in: query
          schema:
            type: ["string", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Computed packages with final prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackageResult"
  /destinations/{slug}/summary:
    get:
      operationId: getDestinationSummary
      tags: [destinations]
      summary: Get destination summary stats (price ranges, hotel counts)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Destination summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationSummary"
...
components:
  schemas:
...
    Destination:
      type: object
      required: [id, slug, nameAr, nameEn, country, heroImage, isActive, isFeatured, sortOrder]
      properties:
        id:
          type: integer
        slug:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        country:
          type: string
        descriptionAr:
          type: ["string", "null"]
        descriptionEn:
          type: ["string", "null"]
        heroImage:
          type: string
        flag:
          type: ["string", "null"]
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        ticketPriceJod:
          type: ["number", "null"]
        hotelCount:
          type: ["integer", "null"]
        minPrice:
          type: ["number", "null"]
...
    DestinationSummary:
      type: object
      required: [slug, totalHotels, totalPackages, minPrice, maxPrice, availableNights, availableAreas, availableMealPlans, availableStars]
      properties:
        slug:
          type: string
        totalHotels:
          type: integer
        totalPackages:
          type: integer
        minPrice:
          type: ["number", "null"]
        maxPrice:
          type: ["number", "null"]
        availableNights:
          type: array
          items:
            type: integer
        availableAreas:
          type: array
          items:
            type: string
        availableMealPlans:
          type: array
          items:
            type: string
        availableStars:
          type: array
          items:
            type: integer
...
    PackageResult:
      type: object
      required: [id, hotelId, hotelNameAr, hotelNameEn, stars, area, mealPlan, roomType, nights, finalPriceJod, isActive]
      properties:
        id:
          type: integer
        hotelId:
          type: integer
        hotelNameAr:
          type: string
        hotelNameEn:
          type: string
        stars:
          type: integer
        area:
          type: ["string", "null"]
        mealPlan:
          type: string
        roomType:
          type: string
        nights:
          type: integer
        finalPriceJod:
          type: number
        finalPriceUsd:
          type: ["number", "null"]
        dateFrom:
          type: ["string", "null"]
        dateTo:
          type: ["string", "null"]
        isActive:
          type: boolean
        hotelImageUrl:
          type: ["string", "null"]
        hotelDescription:
          type: ["string", "null"]
...