# lib/api-zod/src/generated/types/destination.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Al Jood Travel & Tourism API
 * OpenAPI spec version: 0.1.0
 */

export interface Destination {
  id: number;
  slug: string;
  nameAr: string;
  nameEn: string;
  country: string;
  /** @nullable */
  descriptionAr?: string | null;
  /** @nullable */
  descriptionEn?: string | null;
  heroImage: string;
  /** @nullable */
  flag?: string | null;
  isActive: boolean;
  isFeatured: boolean;
  sortOrder: number;
  /** @nullable */
  ticketPriceJod?: number | null;
  /** @nullable */
  hotelCount?: number | null;
  /** @nullable */
  minPrice?: number | null;
}